chore: Update PR based on feedback
