+++ cypress/integration/Popup.spec.ts
@@ -89,4 +89,35 @@ describe('Popup', () => {
       });
     });
   });
+
+  context('given the MultiplePopups story is rendered', () => {
+    beforeEach(() => {
+      h.stories.load('Testing|Popups/Popup/React', 'MultiplePopups');
+    });
+
+    context('when Open Popup 1 button is clicked', () => {
+      beforeEach(() => {
+        cy.findByText('Open Popup 1').click();
+      });
+
+      it('should open Popup 1', () => {
+        cy.findByLabelText('Popup 1').should('be.visible');
+      });
+
+      context('then Open Popup 2 button is click', () => {
+        beforeEach(() => {
+          cy.findByText('Open Popup 2').click();
+        });
+
+        it('should open Popup 2', () => {
+          cy.findByLabelText('Popup 2').should('be.visible');
+        });
+
+        // TODO Skip for now until we have a systematic approach to fix this issue
+        it.skip('should close Popup 1', () => {
+          cy.findAllByLabelText('Popup 1').should('not.exist');
+        });
+      });
+    });
+  });
 });
