+++ cypress/integration/PopupStack.spec.ts
@@ -18,7 +18,7 @@ describe('PopupStack', () => {
   });
 
   beforeEach(() => {
-    h.stories.load('Components|Popups/Popup Stack', 'MixedPopupTypes');
+    h.stories.load('Testing|Popups/Popup Stack', 'MixedPopupTypes');
   });
 
   it('should start with Window 3 stacked on top of 3 Windows', () => {
